Separate elementer for hovedmeny-knapp
