+++ src/komponenter/header/header-regular/common/ekspanderbar-meny/EkspanderbarMeny.tsx
@@ -4,31 +4,27 @@ import './EkspanderbarMeny.less';
 
 type Props = {
     isOpen: boolean;
-    menyKnapp: ReactNode;
     classname: string;
     id: string;
     children: ReactNode;
 };
 
 const EkspanderbarMeny = (props: Props) => {
-    const { isOpen, menyKnapp, classname, id, children } = props;
+    const { isOpen, classname, id, children } = props;
     const cls = BEMHelper('ekspanderbar');
 
     return (
-        <>
-            {menyKnapp}
-            <div className={`${cls.element('container')} ${classname}`}>
-                <div
-                    className={cls.element(
-                        'innhold-wrapper',
-                        isOpen ? 'active' : ''
-                    )}
-                    id={id}
-                >
-                    <div className={cls.element('innhold')}>{children}</div>
-                </div>
+        <div className={`${cls.element('container')} ${classname}`}>
+            <div
+                className={cls.element(
+                    'innhold-wrapper',
+                    isOpen ? 'active' : ''
+                )}
+                id={id}
+            >
+                <div className={cls.element('innhold')}>{children}</div>
             </div>
-        </>
+        </div>
     );
 };
