+++ src/komponenter/header/header-regular/common/knapper/hovedmeny-knapp/HovedmenyKnapp.tsx
@@ -3,26 +3,49 @@ import MenylinjeKnapp from 'komponenter/header/header-regular/common/knapper/Men
 import HamburgerIkon from 'komponenter/header/header-regular/common/knapper/ikoner/hamburger-ikon/HamburgerIkon';
 import Undertittel from 'nav-frontend-typografi/lib/undertittel';
 import Tekst from 'tekster/finn-tekst';
+import { gaEvent } from 'utils/google-analytics';
+import { GACategory } from 'utils/google-analytics';
+import { toggleHovedmeny } from 'store/reducers/dropdown-toggle-duck';
+import { useDispatch } from 'react-redux';
+import { AppState } from 'store/reducers';
+import { useSelector } from 'react-redux';
+
+const stateSelector = (state: AppState) => ({
+    arbeidsflate: state.arbeidsflate.status,
+    isOpen: state.dropdownToggles.hovedmeny,
+});
 
 type Props = {
-    isOpen: boolean;
-    onClick: () => void;
     id?: string;
 };
 
 const classname = 'hovedmeny';
 
-export const HovedmenyKnapp = ({ isOpen, onClick, id }: Props) => (
-    <MenylinjeKnapp
-        isOpen={isOpen}
-        classname={classname}
-        onClick={onClick}
-        ariaControls={classname}
-        id={id}
-    >
-        <HamburgerIkon isOpen={isOpen} />
-        <Undertittel>
-            <Tekst id="meny-knapp" />
-        </Undertittel>
-    </MenylinjeKnapp>
-);
+export const HovedmenyKnapp = ({ id }: Props) => {
+    const dispatch = useDispatch();
+    const { arbeidsflate, isOpen } = useSelector(stateSelector);
+
+    const toggleMenu = () => {
+        gaEvent({
+            context: arbeidsflate,
+            category: GACategory.Header,
+            action: `meny-${isOpen ? 'close' : 'open'}`,
+        });
+        dispatch(toggleHovedmeny());
+    };
+
+    return (
+        <MenylinjeKnapp
+            isOpen={isOpen}
+            classname={classname}
+            onClick={toggleMenu}
+            ariaControls={classname}
+            id={id}
+        >
+            <HamburgerIkon isOpen={isOpen} />
+            <Undertittel>
+                <Tekst id="meny-knapp" />
+            </Undertittel>
+        </MenylinjeKnapp>
+    );
+};
