+++ src/komponenter/header/header-regular/mobil/MobilHeader.less
@@ -8,29 +8,23 @@
     .desktop-sok-dropdown,
     .desktop-minside-meny__knapp,
     .desktop-minside-meny,
-    .logout-knapp {
+    .logout-knapp,
+    #desktop-hovedmeny-knapp-id {
         display: none;
     }
 
     .nav-logo-lenke {
-        order: 1;
         margin-left: @margins-sides-mobil;
         width: 72px;
         padding: 4px;
     }
 
-    &__spacer {
-        order: 2;
-    }
-
     .login-knapp-container,
     .varselbjelle__knapp {
-        order: 3;
         margin-right: 0.5rem;
     }
 
     .hovedmeny__knapp {
-        order: 4;
         margin-right: @margins-sides-mobil;
     }
 }
